Added room creation functionality
Can now create a new room from the homepage, through the '+' icon dialog box. Pressing the create radshare button will make an api call to the database and create a new room in it as well as generate a room code that others will be able to use to join in the future.