Display rooms in database in the table on the homepage
Newly made rooms are now displayed on the home page table and will delete themselves from the database after 2 hours. An expiration date attribute was added to the schema to record the UTC millisecond time that it will be deleted. This time is converted by the client into H:MM format.